feat: implement CI/CD workflow and add cookie inspection tool with CLI and web interface
