+++ cookies.csv
@@ -1,0 +1,6 @@
+Name,Value
+GPS,1
+YSC,_G1_qFsCo_Y
+__Secure-ROLLOUT_TOKEN,CK-MiM3DvJ3CvAEQ79rIpcOZjwMY79rIpcOZjwM%3D
+VISITOR_INFO1_LIVE,KEkisQkI4V4
+VISITOR_PRIVACY_METADATA,CgJJRBIEGgAgbQ%3D%3D
