+++ cookies.json
@@ -1,0 +1,7 @@
+{
+    "GPS": "1",
+    "YSC": "vsffeGLJxhU",
+    "VISITOR_INFO1_LIVE": "-F5Uurh13H0",
+    "VISITOR_PRIVACY_METADATA": "CgJJRBIEGgAgYA%3D%3D",
+    "__Secure-ROLLOUT_TOKEN": "COi1zMK9u_uNCxDTlJ2jw5mPAxjTlJ2jw5mPAw%3D%3D"
+}
