--- cookies_1755553812.json
@@ -1,7 +1,0 @@
-{
-    "GPS": "1",
-    "YSC": "6HxMD9kAt4c",
-    "__Secure-ROLLOUT_TOKEN": "CM3Yv8LJoNTGjAEQ9faoyquVjwMY9faoyquVjwM%3D",
-    "VISITOR_INFO1_LIVE": "Gh6fT5tS3hY",
-    "VISITOR_PRIVACY_METADATA": "CgJJRBIEGgAgYQ%3D%3D"
-}
