+++ pyproject.toml
@@ -1,0 +1,24 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "cookiespy"
+version = "0.1.0"
+description = "A simple cookie inspection & export CLI tool."
+readme = "README.md"
+license = { text = "MIT" }
+authors = [
+  { name="Fadhli Akbar", email="your@email.com" }
+]
+dependencies = [
+  "requests>=2.0",
+]
+requires-python = ">=3.10"
+
+[project.urls]
+Homepage = "https://github.com/fdhliakbar/cookiespy"
+Issues = "https://github.com/fdhliakbar/cookiespy/issues"
+
+[project.scripts]
+cookiespy = "cookiespy.cli:main"
