+++ src/cookiespy.egg-info/dependency_links.txt
@@ -1,0 +1,1 @@
+
