+++ src/cookiespy.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+cookiespy = cookiespy.cli:main
