+++ src/cookiespy.egg-info/requires.txt
@@ -1,0 +1,1 @@
+requests>=2.0
