+++ src/cookiespy/__init__.py
@@ -1,0 +1,4 @@
+from .cli import main
+
+__all__ = ["main"]
+__version__ = "0.1.0"
